\name{appleAll}
\alias{appleAll}
\docType{data}
\title{
Full apple data set
}
\description{
Original USDA PDP data file for the apple data set. The \code{\link{apple}} data set is a derivative of this data set.
}
\usage{data(appleAll)}
\format{
  A data frame with 1137 observations on the following 15 variables.
  \describe{
    \item{\code{pesticide}}{The pesticide measured.}
    \item{\code{sampleID}}{Sample ID.}
    \item{\code{lod}}{Minimum level of detection.}
    \item{\code{comp}}{Measured composite residue.}
    \item{\code{singleWhole}}{Single apple measurement.}
    \item{\code{ss1}}{Single item measurement 1.}
    \item{\code{ss2}}{Single item measurement 2.}
    \item{\code{ss3}}{Single item measurement 3.}
    \item{\code{ss4}}{Single item measurement 4.}
    \item{\code{ss5}}{Single item measurement 5.}
    \item{\code{ss6}}{Single item measurement 6.}
    \item{\code{ss7}}{Single item measurement 7.}
    \item{\code{ss8}}{Single item measurement 8.}
    \item{\code{ss9}}{Single item measurement 9.}
    \item{\code{ss10}}{Single item measurement 10.}
  }
}
\source{
United States Deparatment of Agriculture Pesticide
Data Program.
}
\examples{
data(appleAll)

}
\keyword{datasets}
\keyword{pesticide}
\keyword{commodity}
\keyword{apple}
\keyword{USDA}
